+++ src/main/java/boomcow/minezero/command/SetCheckpointCommand.java
@@ -1,15 +1,14 @@
 package boomcow.minezero.command;
 
-import com.mojang.brigadier.CommandDispatcher;
-import net.minecraft.commands.arguments.EntityArgument;
 import boomcow.minezero.checkpoint.CheckpointManager;
+import com.mojang.brigadier.CommandDispatcher;
 import net.minecraft.commands.CommandSourceStack;
 import net.minecraft.commands.Commands;
 import net.minecraft.commands.arguments.EntityArgument;
 import net.minecraft.network.chat.Component;
 import net.minecraft.server.level.ServerPlayer;
 
-public class SetCheckPointCommand {
+public class SetCheckpointCommand {
     public static void register(CommandDispatcher<CommandSourceStack> dispatcher) {
         dispatcher.register(
                 Commands.literal("setcheckpoint")