+++ src/main/java/boomcow/minezero/event/DeathEventHandler.java
@@ -45,10 +45,10 @@ public class DeathEventHandler {
                 if (!p.getUUID().equals(data.getAnchorPlayerUUID())) {
                 }
             });
-            String chime = ConfigHandler.getDeathChime();
-            if ("CLASSIC".equalsIgnoreCase(chime)) {
+            String restoreSound = ConfigHandler.getRestoreSound();
+            if ("death_chime".equalsIgnoreCase(restoreSound) || "CLASSIC".equalsIgnoreCase(restoreSound)) {
                 playClassicChime(player);
-            } else if ("ALTERNATE".equalsIgnoreCase(chime)) {
+            } else if ("alt_death_chime".equalsIgnoreCase(restoreSound) || "ALTERNATE".equalsIgnoreCase(restoreSound)) {
                 playAlternateChime(player);
             }